requirements.txt
streamlit==1.40.1
langchain==1.1.0
langchain-core==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-chroma==1.0.0
langchain-huggingface==1.1.0
langchain-groq==1.1.0
chromadb==1.3.5
sentence-transformers==5.1.2
python-dotenv==1.0.1
requests==2.32.5
